+++ apps/explorer/lib/explorer/celo/events/contract_events/election/validator_group_vote_activated_event.ex
@@ -10,6 +10,10 @@ defmodule Explorer.Celo.ContractEvents.Election.ValidatorGroupVoteActivatedEvent
     );
   """
 
+  alias Explorer.Celo.ContractEvents.Common
+  alias Explorer.Chain.{Block, CeloContractEvent}
+  alias Explorer.Repo
+
   use Explorer.Celo.ContractEvents.Base,
     name: "ValidatorGroupVoteActivated",
     topic: "0x45aac85f38083b18efe2d441a65b9c1ae177c78307cb5a5d4aec8f7dbcaeabfe"
@@ -18,4 +22,21 @@ defmodule Explorer.Celo.ContractEvents.Election.ValidatorGroupVoteActivatedEvent
   event_param(:group, :address, :indexed)
   event_param(:value, {:uint, 256}, :unindexed)
   event_param(:units, {:uint, 256}, :unindexed)
+
+  def get_account_group_pairs_with_activated_votes(block_number) do
+    query =
+      from(
+        event in CeloContractEvent,
+        where: event.name == "ValidatorGroupVoteActivated",
+        where: event.block_number < ^block_number,
+        select: %{
+          account_hash: json_extract_path(event.params, ["account"]),
+          group_hash: json_extract_path(event.params, ["group"])
+        }
+      )
+
+    query
+    |> Repo.all()
+    |> Enum.map(&%{account_hash: Common.ca(&1.account_hash), group_hash: Common.ca(&1.group_hash)})
+  end
 end