+++ apps/explorer/lib/explorer/chain.ex
@@ -2863,6 +2863,24 @@ defmodule Explorer.Chain do
     Repo.stream_reduce(query, initial, reducer)
   end
 
+  @spec stream_blocks_with_unfetched_voter_votes(
+          initial :: accumulator,
+          reducer :: (entry :: term(), accumulator -> accumulator)
+        ) :: {:ok, accumulator}
+        when accumulator: term()
+  def stream_blocks_with_unfetched_voter_votes(initial, reducer) when is_function(reducer, 2) do
+    query =
+      from(
+        b in Block,
+        join: celo_pending_ops in assoc(b, :celo_pending_epoch_operations),
+        where: celo_pending_ops.fetch_voter_votes,
+        select: %{block_number: b.number, block_hash: b.hash},
+        order_by: [asc: b.number]
+      )
+
+    Repo.stream_reduce(query, initial, reducer)
+  end
+
   def remove_nonconsensus_blocks_from_pending_ops(block_hashes) do
     query =
       from(