+++ apps/explorer/lib/explorer/chain/import/runner/celo_epoch_rewards.ex
@@ -43,10 +43,10 @@ defmodule Explorer.Chain.Import.Runner.CeloEpochRewards do
       end)
 
     multi_chain
-    |> Multi.run(:delete_celo_pending, fn _, _ ->
+    |> Multi.run(:falsify_fetch_epoch_rewards, fn _, _ ->
       changes_list
       |> Enum.each(fn reward ->
-        CeloPendingEpochOperation.falsify_or_delete_celo_pending_epoch_operation(
+        CeloPendingEpochOperation.falsify_celo_pending_epoch_operation(
           reward.block_hash,
           :fetch_epoch_rewards
         )