+++ apps/explorer/priv/repo/migrations/20220312200304_add_initial_celo_pending_operations_voter_votes.exs
@@ -1,0 +1,19 @@
+defmodule Explorer.Repo.Migrations.AddInitialCeloPendingEpochOperationsVoterVotes do
+  use Ecto.Migration
+
+  def up do
+    execute("""
+    WITH epoch_blocks AS (
+    SELECT i * 17280 as block_number FROM generate_series(1, (SELECT (MAX(number)/17280) FROM blocks)) as i
+    ), epoch_block_hashes AS
+    (SELECT b.hash, true as epoch, (CASE WHEN b.number < 172800 THEN false ELSE true END) as validator_group, true as voter_votes FROM epoch_blocks eb LEFT JOIN blocks b ON b.number = eb.block_number where b.hash is not null)
+    INSERT INTO celo_pending_epoch_operations (
+    block_hash, fetch_epoch_rewards, fetch_validator_group_data, fetch_voter_votes, inserted_at, updated_at
+    ) SELECT *, NOW(), NOW() FROM epoch_block_hashes;
+    """)
+  end
+
+  def down do
+    execute("delete from celo_pending_epoch_operations;")
+  end
+end