+++ apps/indexer/test/indexer/fetcher/celo_validator_group_votes_test.exs
@@ -142,7 +142,6 @@ defmodule Indexer.Fetcher.CeloValidatorGroupVotesTest do
 
       CeloValidatorGroupVotesFetcher.import_items(votes)
 
-      assert count(CeloPendingEpochOperation) == 1
       assert count(CeloValidatorGroupVotes) == 2
     end
   end